Deprecates CityFingerprint32 because it's slower and more prone to hash collisions than the 64-bit version.
PiperOrigin-RevId: 808903617

+++ src/composer/table.h
@@ -176,7 +176,7 @@ class TableManager {
   //  config::Config::SymbolMethod
   absl::flat_hash_map<size_t, std::shared_ptr<const Table>> table_map_;
   // Fingerprint for Config::custom_roman_table;
-  uint32_t custom_roman_table_fingerprint_;
+  uint64_t custom_roman_table_fingerprint_;
 };
 
 }  // namespace composer
